+++ pyuploadcare/conf.py
@@ -5,26 +5,32 @@ Configuration variables.
 """
 import os
 
-from pyuploadcare.api.utils import get_cname_prefix
+from pyuploadcare.helpers import get_cdn_base
 
 # Do not import vars from this module.
 # Instead import whole module and work with attributes.
 
 __all__ = []
 
+
 DEFAULT = object()
+DEFAULT_API_BASE = "https://api.uploadcare.com/"
+DEFAULT_UPLOAD_BASE = "https://upload.uploadcare.com/"
+DEFAULT_CDN_BASE = "https://ucarecdn.com/"
+DEFAULT_USE_SUBDOMAINS = "no"
 
 pub_key = os.getenv("UPLOADCARE_PUBLIC_KEY")
 secret = os.getenv("UPLOADCARE_SECRET_KEY")
 
 api_version = "0.7"
-api_base = os.getenv("UPLOADCARE_API_BASE", "https://api.uploadcare.com/")
-upload_base = os.getenv(
-    "UPLOADCARE_UPLOAD_BASE", "https://upload.uploadcare.com/"
-)
-prefix = get_cname_prefix(pub_key)
-cdn_base = os.getenv("UPLOADCARE_CDN_BASE",
-                     f"https://{prefix}.ucarecd.net/")
+api_base = os.getenv("UPLOADCARE_API_BASE", DEFAULT_API_BASE)
+upload_base = os.getenv("UPLOADCARE_UPLOAD_BASE", DEFAULT_UPLOAD_BASE)
+
+use_subdomains = os.getenv("UPLOADCARE_USE_SUBDOMAINS",
+                           DEFAULT_USE_SUBDOMAINS).lower() in ("true", "yes")
+cdn_base = get_cdn_base(pub_key,
+                        default=DEFAULT_CDN_BASE,
+                        subdomains=use_subdomains)
 
 signed_uploads = True
 signed_uploads_ttl = 60